Fix CPU load calculation overflow with DWT->CYCCNT counter
idle_delta * 100UL overflows uint32_t when the counter runs at 250 MHz
(DWT->CYCCNT). At 96% idle time, idle_delta ≈ 240M cycles; ×100 = 24B
which wraps to ~88% instead of the correct ~4% CPU load.

Fix: cast to uint64_t before the multiply.

+++ code/GO_board.c
@@ -357,7 +357,7 @@ static void ControllerInfoTask(void *args) {
 							                      - prev_idle_time;
 							prev_idle_time = s_task_status[i].ulRunTimeCounter;
 							go_board_cpu_load = (uint8_t)(100u
-							    - (uint8_t)((idle_delta * 100UL) / total_delta));
+							    - (uint8_t)(((uint64_t)idle_delta * 100ULL) / total_delta));
 							break;
 						}
 					}
